Centralized home services on single file. Changed imports in all home components
